Update Advancement index.md
Reorganized the Advancement table (BECAUSE IT WAS DRIVING ME INSANE xD) and added back Armor Proficiency

Maybe consider removing Purchasing Additional Spells and just put it as the front page of the Spells page? https://stormchaserroleplaying.com/stormchaserRPG/Spells/

Maybe the same for Attributes? It's a little confusing right now when you try to compare it to the Generating Attributes table